Small re-factor cleanup
