+++ AutoDI.Fody/ModuleWeaver - Copy.InjectContainer.cs
@@ -1,0 +1,5 @@
+﻿// ReSharper disable once CheckNamespace
+partial class ModuleWeaver
+{
+    
+}
