+++ AutoDI.Fody/ModuleWeaver.InjectContainer
@@ -1,0 +1,4 @@
+﻿namespace NAMESPACE
+{
+  public class CLASS {}
+}
