+++ AutoDI.Fody/ModuleWeaver.cs
@@ -14,7 +14,7 @@ using OpCodes = Mono.Cecil.Cil.OpCodes;
 
 [assembly: InternalsVisibleTo("AutoDI.Fody.Tests")]
 // ReSharper disable once CheckNamespace
-public class ModuleWeaver
+public partial class ModuleWeaver
 {
     // Will contain the full element XML from FodyWeavers.xml. OPTIONAL
     public XElement Config { get; set; }
@@ -85,9 +85,12 @@ public class ModuleWeaver
                 {
                     LogDebug(s);
                 }
+
             };
 
-            Mapping mapping = GetMapping(settings);
+            ICollection<TypeDefinition> allTypes = GetAllTypes(settings);
+            Mapping mapping = GetMapping(settings, allTypes);
+
             InternalLogDebug($"Found potential map:\r\n{mapping}", DebugLogLevel.Verbose);
             TypeDefinition resolverType = CreateAutoDIContainer(mapping);
 
@@ -98,7 +101,7 @@ public class ModuleWeaver
                 InjectContainer(resolverType);
             }
 
-            foreach (TypeDefinition type in ModuleDefinition.GetAllTypes())
+            foreach (TypeDefinition type in allTypes)
             {
                 ProcessType(type);
             }
@@ -112,59 +115,17 @@ public class ModuleWeaver
         }
     }
 
-    private void InjectContainer(TypeDefinition resolverType)
-    {
-        if (ModuleDefinition.EntryPoint != null)
-        {
-            ILProcessor entryMethodProcessor = ModuleDefinition.EntryPoint.Body.GetILProcessor();
-            var create = Instruction.Create(OpCodes.Newobj,
-                resolverType.Methods.Single(m => m.IsConstructor && !m.IsStatic));
-            var setMethod = ModuleDefinition.ImportReference(typeof(DependencyResolver).GetMethod(
-                nameof(DependencyResolver.Set),
-                new[] { typeof(IDependencyResolver) }));
-            var set = Instruction.Create(OpCodes.Call, setMethod);
-            entryMethodProcessor.InsertBefore(ModuleDefinition.EntryPoint.Body.Instructions.First(), set);
-            entryMethodProcessor.InsertBefore(set, create);
-        }
-        else
-        {
-            InternalLogDebug($"No entry point in {ModuleDefinition.FileName}. Skipping container injection.", DebugLogLevel.Default);
-        }
-    }
-
-    private Mapping GetMapping(Settings settings)
-    {
-        var rv = new Mapping();
-        ICollection<TypeDefinition> allTypes = GetAllTypes(settings);
-
-        if (settings.Behavior.HasFlag(Behaviors.SingleInterfaceImplementation))
-        {
-            AddSingleInterfaceImplementation(rv, allTypes);
-        }
-        if (settings.Behavior.HasFlag(Behaviors.IncludeClasses))
-        {
-            AddClasses(rv, allTypes);
-        }
-        if (settings.Behavior.HasFlag(Behaviors.IncludeBaseClasses))
-        {
-            AddBaseClasses(rv, allTypes);
-        }
-
-        AddSettingsMap(settings, rv, allTypes);
-
-        return rv;
-    }
-
     private ICollection<TypeDefinition> GetAllTypes(Settings settings)
     {
         var allTypes = new HashSet<TypeDefinition>(TypeComparer.FullName);
-        IEnumerable<TypeDefinition> FilterTypes(IEnumerable<TypeDefinition> types) => 
+        IEnumerable<TypeDefinition> FilterTypes(IEnumerable<TypeDefinition> types) =>
             types.Where(t => !t.IsCompilerGenerated() && !allTypes.Remove(t));
 
         foreach (TypeDefinition type in FilterTypes(ModuleDefinition.GetAllTypes()))
         {
             allTypes.Add(type);
         }
+
         foreach (AssemblyNameReference assemblyReference in ModuleDefinition.AssemblyReferences)
         {
             bool useAutoDiAssebmlies = settings.Behavior.HasFlag(Behaviors.IncludeDependentAutoDIAssemblies);
@@ -173,7 +134,7 @@ public class ModuleWeaver
             {
                 AssemblyDefinition assembly = AssemblyResolver.Resolve(assemblyReference);
                 if (assembly == null) continue;
-                
+
                 //Check if it references AutoDI. If it doesn't we will skip
                 if (!matchesAssembly && assembly.MainModule.AssemblyReferences.All(a =>
                         a.FullName != typeof(DependencyAttribute).Assembly.FullName))
@@ -190,372 +151,6 @@ public class ModuleWeaver
         return allTypes;
     }
 
-    private static void AddSingleInterfaceImplementation(Mapping map, IEnumerable<TypeDefinition> allTypes)
-    {
-        var types = new Dictionary<TypeReference, List<TypeDefinition>>(TypeComparer.FullName);
-
-        foreach (TypeDefinition type in allTypes.Where(t => t.IsClass && !t.IsAbstract))
-        {
-            foreach (var @interface in type.Interfaces)
-            {
-                if (!types.TryGetValue(@interface.InterfaceType, out List<TypeDefinition> list))
-                {
-                    types.Add(@interface.InterfaceType, list = new List<TypeDefinition>());
-                }
-                list.Add(type);
-                //TODO: Base types
-            }
-        }
-
-        foreach (KeyValuePair<TypeReference, List<TypeDefinition>> kvp in types)
-        {
-            if (kvp.Value.Count != 1) continue;
-            map.Add(kvp.Key.Resolve(), kvp.Value[0], DuplicateKeyBehavior.RemoveAll);
-        }
-    }
-
-    private static void AddClasses(Mapping map, IEnumerable<TypeDefinition> types)
-    {
-        foreach (TypeDefinition type in types.Where(t => t.IsClass && !t.IsAbstract))
-        {
-            map.Add(type, type, DuplicateKeyBehavior.RemoveAll);
-        }
-    }
-
-    private static void AddBaseClasses(Mapping map, IEnumerable<TypeDefinition> types)
-    {
-        TypeDefinition GetBaseType(TypeDefinition type)
-        {
-            return type.BaseType?.Resolve();
-        }
-
-        foreach (TypeDefinition type in types.Where(t => t.IsClass && !t.IsAbstract && t.BaseType != null))
-        {
-            for (TypeDefinition t = GetBaseType(type); t != null; t = GetBaseType(t))
-            {
-                if (t.FullName != typeof(object).FullName)
-                {
-                    map.Add(t, type, DuplicateKeyBehavior.RemoveAll);
-                }
-            }
-        }
-    }
-
-    private void AddSettingsMap(Settings settings, Mapping map, IEnumerable<TypeDefinition> types)
-    {
-        var allTypes = types.ToDictionary(x => x.FullName);
-
-        foreach (string typeName in allTypes.Keys)
-        {
-            foreach (Map settingsMap in settings.Maps)
-            {
-                //TODO: Logging for the various cases where we don't map...
-                if (settingsMap.TryGetMap(typeName, out string mappedType) &&
-                    allTypes.TryGetValue(mappedType, out TypeDefinition mapped) &&
-                    (settingsMap.Force || CanBeCastToType(allTypes[typeName], mapped)))
-                {
-                    map.Add(allTypes[typeName], mapped, DuplicateKeyBehavior.Replace);
-                }
-            }
-        }
-        map.UpdateCreation(settings.Types);
-    }
-
-    private bool CanBeCastToType(TypeDefinition key, TypeDefinition targetType)
-    {
-        var comparer = TypeComparer.FullName;
-
-        for (TypeDefinition t = targetType; t != null; t = t.BaseType?.Resolve())
-        {
-            if (comparer.Equals(key, t)) return true;
-            if (t.Interfaces.Any(i => comparer.Equals(i.InterfaceType, key)))
-            {
-                return true;
-            }
-        }
-        InternalLogDebug($"'{targetType.FullName}' cannot be cast to '{key.FullName}', ignoring", DebugLogLevel.Verbose);
-        return false;
-    }
-
-    private TypeDefinition CreateAutoDIContainer(Mapping mapping)
-    {
-        var type = new TypeDefinition("AutoDI", "AutoDIContainer",
-            TypeAttributes.Class | TypeAttributes.Public)
-        {
-            BaseType = ModuleDefinition.Get<object>()
-        };
-        MethodDefinition ctor = CreateContainerConstructor();
-        type.Methods.Add(ctor);
-
-
-        //Create static constructor
-        MethodDefinition staticConstructor = CreateStaticConstructor();
-        ILProcessor staticBody = staticConstructor.Body.GetILProcessor();
-
-        //Declare and initialize dictionary map
-        FieldDefinition mapField = CreateStaticReadonlyField<ContainerMap>("_items", false);
-        type.Fields.Add(mapField);
-        MethodReference mapConstructor = ModuleDefinition.ImportReference(typeof(ContainerMap).GetConstructor(new Type[0]));
-        staticBody.Emit(OpCodes.Newobj, mapConstructor);
-        staticBody.Emit(OpCodes.Stsfld, mapField);
-
-        BuildMap(mapField, staticBody, type, mapping);
-
-        //Pass map to setup method if one exists
-        MethodDefinition setupMethod = FindSetupMethod();
-        if (setupMethod != null)
-        {
-            InternalLogDebug($"Found setup method '{setupMethod.FullName}'", DebugLogLevel.Verbose);
-            staticBody.Emit(OpCodes.Ldsfld, mapField);
-            staticBody.Emit(OpCodes.Call, setupMethod);
-            staticBody.Emit(OpCodes.Nop);
-        }
-
-        staticBody.Emit(OpCodes.Ret);
-
-        type.Methods.Add(staticConstructor);
-
-        //Implement IDependencyResolver interface
-        TypeReference dependencyResolver = ModuleDefinition.Get<IDependencyResolver>();
-        type.Interfaces.Add(new InterfaceImplementation(dependencyResolver));
-
-        //Create resolve method
-        var resolveMethod = new MethodDefinition("Resolve",
-            MethodAttributes.Private | MethodAttributes.Final | MethodAttributes.HideBySig | MethodAttributes.Virtual |
-            MethodAttributes.NewSlot
-            , type);
-        resolveMethod.Parameters.Add(new ParameterDefinition("parameters", ParameterAttributes.None, ModuleDefinition.Get<object[]>()));
-
-        resolveMethod.Overrides.Add(ModuleDefinition.ImportReference(dependencyResolver.Resolve().Methods.Single()));
-
-        var genericParameter = new GenericParameter("T", resolveMethod);
-        resolveMethod.GenericParameters.Add(genericParameter);
-        resolveMethod.ReturnType = genericParameter;
-
-        var lazy = new VariableDefinition(ModuleDefinition.Get<Lazy<object>>());
-        resolveMethod.Body.Variables.Add(lazy);
-        var genericVariable = new VariableDefinition(genericParameter);
-        resolveMethod.Body.Variables.Add(genericVariable);
-
-        var body = resolveMethod.Body.GetILProcessor();
-        body.Emit(OpCodes.Ldsfld, mapField);
-
-        MethodReference getMethod =
-            ModuleDefinition.ImportReference(typeof(ContainerMap).GetMethod(nameof(ContainerMap.Get)));
-        var genericGetMethod = new GenericInstanceMethod(getMethod);
-        genericGetMethod.GenericArguments.Add(genericParameter);
-
-        body.Emit(OpCodes.Call, genericGetMethod);
-        body.Emit(OpCodes.Ret);
-
-        type.Methods.Add(resolveMethod);
-
-        return type;
-    }
-
-    private void BuildMap(FieldDefinition mapField, ILProcessor staticBody, TypeDefinition delegateContainer, Mapping mapping)
-    {
-        bool InvokeConstructor(TypeDefinition targetType, ILProcessor processor)
-        {
-            foreach (MethodDefinition targetTypeCtor in targetType.GetConstructors().OrderByDescending(c => c.Parameters.Count))
-            {
-                if (targetTypeCtor.Parameters.All(pd => pd.HasDefault && pd.Constant == null))
-                {
-                    for (int i = 0; i < targetTypeCtor.Parameters.Count; i++)
-                    {
-                        processor.Emit(OpCodes.Ldnull);
-                    }
-                    processor.Emit(OpCodes.Newobj, ModuleDefinition.ImportReference(targetTypeCtor));
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        MethodReference addSingleton =
-            ModuleDefinition.ImportReference(typeof(ContainerMap).GetMethod(nameof(ContainerMap.AddSingleton)));
-        MethodReference addLazySingleton =
-            ModuleDefinition.ImportReference(typeof(ContainerMap).GetMethod(nameof(ContainerMap.AddLazySingleton)));
-        MethodReference addWeakTransient =
-            ModuleDefinition.ImportReference(typeof(ContainerMap).GetMethod(nameof(ContainerMap.AddWeakTransient)));
-        MethodReference addTransient =
-            ModuleDefinition.ImportReference(typeof(ContainerMap).GetMethod(nameof(ContainerMap.AddTransient)));
-
-        TypeReference funcType = ModuleDefinition.ImportReference(typeof(Func<>));
-        //NB: This null fall back is due to an issue with Mono.Cecil 0.10.0
-        //From GitHub issues it looks like it make be resolved in some of the beta builds, however this would require modifying Fody.
-        //For now we will just manually resolve this way. Since we know Func<>> lives in the core library.
-        TypeDefinition funcDefinition = funcType.Resolve() ?? ModuleDefinition.AssemblyResolver
-                                            .Resolve((AssemblyNameReference)ModuleDefinition.TypeSystem.CoreLibrary)
-                                            .MainModule.GetType(typeof(Func<>).FullName);
-        if (funcDefinition == null)
-        {
-            LogError($"Failed to resolve type '{typeof(Func<>).FullName}'");
-            return;
-        }
-        MethodDefinition funcCtor = funcDefinition.GetConstructors().Single();
-
-        TypeReference type = ModuleDefinition.ImportReference(typeof(Type));
-        MethodReference getTypeMethod = ModuleDefinition.ImportReference(typeof(Type).GetMethod(nameof(Type.GetTypeFromHandle)));
-
-        int delegateMethodCount = 0;
-        foreach (TypeMap map in mapping)
-        {
-            try
-            {
-                InternalLogDebug($"Processing map for {map.TargetType.FullName}", DebugLogLevel.Verbose);
-
-                TypeDefinition targetType = map.TargetType;
-                staticBody.Emit(OpCodes.Ldsfld, mapField);
-
-                switch (map.Lifetime)
-                {
-                    case Lifetime.Singleton:
-                        if (!InvokeConstructor(targetType, staticBody))
-                        {
-                            staticBody.Remove(staticBody.Body.Instructions.Last());
-                            InternalLogDebug($"No acceptable constructor for '{targetType.FullName}', skipping map", DebugLogLevel.Verbose);
-                            continue;
-                        }
-                        break;
-                    default:
-                        var delegateMethod = new MethodDefinition($"<{targetType.Name}>_generated_{delegateMethodCount++}",
-                            MethodAttributes.Assembly | MethodAttributes.HideBySig | MethodAttributes.Static |
-                            MethodAttributes.Private, ModuleDefinition.ImportReference(targetType));
-                        ILProcessor delegateProcessor = delegateMethod.Body.GetILProcessor();
-                        if (!InvokeConstructor(targetType, delegateProcessor))
-                        {
-                            delegateMethodCount--;
-                            staticBody.Remove(staticBody.Body.Instructions.Last());
-                            InternalLogDebug($"No acceptable constructor for '{targetType.FullName}', skipping map", DebugLogLevel.Verbose);
-                            continue;
-                        }
-                        delegateProcessor.Emit(OpCodes.Ret);
-                        delegateContainer.Methods.Add(delegateMethod);
-
-                        staticBody.Emit(OpCodes.Ldnull);
-                        staticBody.Emit(OpCodes.Ldftn, delegateMethod);
-
-                        staticBody.Emit(OpCodes.Newobj, ModuleDefinition.ImportReference(GetGenericTypeConstructor(funcCtor, targetType)));
-                        break;
-                }
-
-                staticBody.Emit(OpCodes.Ldc_I4, map.Keys.Count);
-                staticBody.Emit(OpCodes.Newarr, type);
-
-                int arrayIndex = 0;
-                foreach (TypeDefinition key in map.Keys)
-                {
-                    TypeReference importedKey = ModuleDefinition.ImportReference(key);
-                    InternalLogDebug($"Mapping {importedKey.FullName} => {targetType.FullName} ({map.Lifetime})", DebugLogLevel.Default);
-                    staticBody.Emit(OpCodes.Dup);
-                    staticBody.Emit(OpCodes.Ldc_I4, arrayIndex++);
-                    staticBody.Emit(OpCodes.Ldtoken, importedKey);
-                    staticBody.Emit(OpCodes.Call, getTypeMethod);
-                    staticBody.Emit(OpCodes.Stelem_Ref);
-                }
-
-                MethodReference addMethod;
-                switch (map.Lifetime)
-                {
-                    case Lifetime.Singleton:
-                        addMethod = addSingleton;
-                        break;
-                    case Lifetime.LazySingleton:
-                        addMethod = addLazySingleton;
-                        break;
-                    case Lifetime.WeakTransient:
-                        addMethod = addWeakTransient;
-                        break;
-                    case Lifetime.Transient:
-                        addMethod = addTransient;
-                        break;
-                    default:
-                        throw new InvalidOperationException($"Invalid Crate value '{map.Lifetime}'");
-                }
-                var addGenericMethod = new GenericInstanceMethod(addMethod);
-                addGenericMethod.GenericArguments.Add(ModuleDefinition.ImportReference(targetType));
-
-                staticBody.Emit(OpCodes.Call, addGenericMethod);
-                staticBody.Emit(OpCodes.Nop);
-            }
-            catch (Exception e)
-            {
-                LogWarning($"Failed to create map for {map}\r\n{e}");
-            }
-        }
-    }
-
-    //Based on example from here: https://stackoverflow.com/questions/16430947/emit-call-to-system-lazyt-constructor-with-mono-cecil
-    public static MethodReference GetGenericTypeConstructor(MethodReference self, params TypeReference[] args)
-    {
-        var reference = new MethodReference(
-            self.Name,
-            self.ReturnType,
-            self.DeclaringType.MakeGenericInstanceType(args))
-        {
-            HasThis = self.HasThis,
-            ExplicitThis = self.ExplicitThis,
-            CallingConvention = self.CallingConvention
-        };
-
-        foreach (var parameter in self.Parameters)
-        {
-            reference.Parameters.Add(new ParameterDefinition(parameter.ParameterType));
-        }
-
-        foreach (var genericParam in self.GenericParameters)
-        {
-            reference.GenericParameters.Add(new GenericParameter(genericParam.Name, reference));
-        }
-
-        return reference;
-    }
-
-    private MethodDefinition CreateContainerConstructor()
-    {
-        var ctor = new MethodDefinition(".ctor",
-            MethodAttributes.Public | MethodAttributes.HideBySig | MethodAttributes.SpecialName |
-            MethodAttributes.RTSpecialName, ModuleDefinition.ImportReference(typeof(void)));
-        ILProcessor processor = ctor.Body.GetILProcessor();
-        processor.Emit(OpCodes.Ldarg_0); //this
-        processor.Emit(OpCodes.Call, ModuleDefinition.ImportReference(typeof(object).GetConstructor(new Type[0])));
-        processor.Emit(OpCodes.Ret);
-        return ctor;
-    }
-
-    private MethodDefinition CreateStaticConstructor()
-    {
-        var ctor = new MethodDefinition(".cctor",
-            MethodAttributes.Private | MethodAttributes.HideBySig | MethodAttributes.Static |
-            MethodAttributes.SpecialName | MethodAttributes.RTSpecialName, ModuleDefinition.ImportReference(typeof(void)));
-
-        return ctor;
-    }
-
-    private FieldDefinition CreateStaticReadonlyField<T>(string name, bool @public)
-    {
-        return CreateStaticReadonlyField(name, @public, ModuleDefinition.Get<T>());
-    }
-
-    private static FieldDefinition CreateStaticReadonlyField(string name, bool @public, TypeReference type)
-    {
-        return new FieldDefinition(name,
-            (@public ? FieldAttributes.Public : FieldAttributes.Private) | FieldAttributes.Static |
-            FieldAttributes.InitOnly, type);
-    }
-
-    private MethodDefinition FindSetupMethod()
-    {
-        return ModuleDefinition
-            .GetAllTypes()
-            .SelectMany(t => t.GetMethods())
-            .FirstOrDefault(md => md.IsStatic &&
-                                  (md.IsPublic || md.IsAssembly) &&
-                                  md.CustomAttributes.Any(a => a.AttributeType.IsType<SetupMethodAttribute>()) &&
-                                  md.Parameters.Count == 1 &&
-                                  md.Parameters[0].ParameterType.IsType<ContainerMap>());
-    }
 
     private void ProcessType(TypeDefinition type)
     {
@@ -766,7 +361,6 @@ public class ModuleWeaver
         }
         LogWarning($"Unknown constant type {constant.GetType().FullName}");
     }
-
 
     // Will be called when a request to cancel the build occurs. OPTIONAL
     public void Cancel()
